+++ src/main/java/baseball/Application.java
@@ -1,7 +1,5 @@
 package baseball;
 
-import baseball.domain.GameController;
-
 public class Application {
     public static void main(String[] args) {
         GameController gameController = new GameController();
