+++ src/main/java/baseball/Computer.java
@@ -1,4 +1,4 @@
-package baseball.domain;
+package baseball;
 
 import static camp.nextstep.edu.missionutils.Randoms.pickNumberInRange;
