+++ src/main/java/baseball/GameController.java
@@ -1,4 +1,4 @@
-package baseball.domain;
+package baseball;
 
 import static camp.nextstep.edu.missionutils.Console.readLine;
