+++ src/main/java/baseball/Player.java
@@ -1,4 +1,4 @@
-package baseball.domain;
+package baseball;
 
 import java.util.ArrayList;
 import java.util.List;
