+++ src/test/java/baseball/InputValidationTest.java
@@ -1,7 +1,8 @@
-package baseball.validation;
+package baseball;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import baseball.InputValidation;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
