+++ src/test/java/baseball/JudgementTest.java
@@ -1,7 +1,8 @@
-package baseball.domain;
+package baseball;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import baseball.Judgement;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
