+++ src/test/java/baseball/PlayerTest.java
@@ -1,5 +1,6 @@
-package baseball.domain;
+package baseball;
 
+import baseball.Player;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
